Bump rand to 0.9.2

+++ rs/src/library_chcker/dynamic_sequence_range_affine_range_sum.rs
@@ -371,7 +371,7 @@ where
     left: Option<Box<TreapNode<K, M, Op>>>,
     right: Option<Box<TreapNode<K, M, Op>>>,
     size: usize,
-    priority: usize,
+    priority: u64,
     key: K,
     reversed: bool,
     lazy: Op::Item,
@@ -387,7 +387,7 @@ where
     M: Monoid,
     Op: Monoid + Operator<M::Item>,
 {
-    fn new(priority: usize, key: K, value: M::Item) -> TreapNode<K, M, Op> {
+    fn new(priority: u64, key: K, value: M::Item) -> TreapNode<K, M, Op> {
         TreapNode {
             left: None,
             right: None,
@@ -543,7 +543,7 @@ where
         }
     }
 
-    fn insert(t: Option<Box<Self>>, priority: usize, key: K, value: M::Item) -> Box<Self> {
+    fn insert(t: Option<Box<Self>>, priority: u64, key: K, value: M::Item) -> Box<Self> {
         let mut t = if let Some(t) = t {
             t
         } else {
@@ -575,7 +575,7 @@ where
     fn insert_at(
         t: Option<Box<Self>>,
         nth: usize,
-        priority: usize,
+        priority: u64,
         key: K,
         value: M::Item,
     ) -> Box<Self> {
@@ -752,7 +752,7 @@ where
         self.root = Some(TreapNode::insert_at(
             std::mem::replace(&mut self.root, None),
             nth,
-            self.rng.borrow_mut().next() as usize,
+            self.rng.borrow_mut().next() as u64,
             (),
             value,
         ));
